requirements.txt
python-dotenv
pytz
tabulate
types-tabulate
colorama
types-colorama